[#26] refactor: 테스트 코드 위치 변경
- 코드 커버리지 확인을 위해 해당 코드와 같은 위치로 이동합니다.
